requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Flask-Caching
WTForms-Alchemy
bcrypt
psycopg2
python-slugify
email-validator
phonenumbers
phone-number-validator
unidecode
pandas
valuta